Initializing hal_init() to preserve resource creation order
As part of resource creation, there is a possibility that some of the
resources can be created as Auto Observable, but that will require that
those resources have certain base component like Timer to be created.

Timer creation requires that event scheduler must have been created before hand.

Since, Resource creation is an independent operation than instantiating Pelion Client,
there is a chance that Resource can be created before Client stack is instantiated so it
becomes highly dependant on order in which APIs are called.

To resolve this issue, scheduler initilization call is also added in M2MBase constructor
as fail-safe mechanism, so that order of API calls does not matter for application developer.